+++ utils/constants.go
@@ -5,10 +5,10 @@ import (
 )
 
 const (
-	// The amount of fees to be sent with a default transaction.
+	// defaultFees is the amount of fees to be sent with a default transaction.
 	defaultFees int = 1e18 // 1 aevmos
-	// The amount of blocks in the future that the upgrade will be scheduled.
+	// DeltaHeight is the amount of blocks in the future that the upgrade will be scheduled.
 	DeltaHeight = 10
-	// The denomination used for the local node.
+	// denom is the denomination used for the local node.
 	denom = evmosutils.BaseDenom
 )